agent/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
openai==1.14.2
httpx==0.27.0
python-dotenv==1.0.1
solders==0.21.0
solana>=0.30.2
